requirements.txt
jsonschema==4.23.0
pandas==2.2.3
pandasql==0.7.3
pymongo==4.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
plotly==5.24.1
PyYAML~=6.0.2
streamlit~=1.40.1
pydantic~=2.10.4